backend/pyproject.toml
[project]
name = "savor-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy-utils>=0.42.1",
]


[tool.pytest.ini_options]
pythonpath = [
  "."
]
